+++ src/cdk/a11y/public_api.ts
@@ -12,12 +12,13 @@ import {LIVE_ANNOUNCER_PROVIDER} from './live-announcer';
 import {InteractivityChecker} from './interactivity-checker';
 import {CommonModule} from '@angular/common';
 import {PlatformModule} from '@angular/cdk/platform';
+import {AriaDescriber} from './aria-describer';
 
 @NgModule({
   imports: [CommonModule, PlatformModule],
   declarations: [FocusTrapDirective, FocusTrapDeprecatedDirective],
   exports: [FocusTrapDirective, FocusTrapDeprecatedDirective],
-  providers: [InteractivityChecker, FocusTrapFactory, LIVE_ANNOUNCER_PROVIDER]
+  providers: [InteractivityChecker, FocusTrapFactory, AriaDescriber, LIVE_ANNOUNCER_PROVIDER]
 })
 export class A11yModule {}
 
@@ -28,3 +29,4 @@ export * from './interactivity-checker';
 export * from './list-key-manager';
 export * from './activedescendant-key-manager';
 export * from './focus-key-manager';
+export * from './aria-describer';
