+++ src/lib/tooltip/index.ts
@@ -12,6 +12,7 @@ import {OverlayModule} from '@angular/cdk/overlay';
 import {PlatformModule} from '@angular/cdk/platform';
 import {MdCommonModule} from '../core';
 import {MdTooltip, TooltipComponent, MD_TOOLTIP_SCROLL_STRATEGY_PROVIDER} from './tooltip';
+import {A11yModule} from '@angular/cdk/a11y';
 
 
 @NgModule({
@@ -19,7 +20,8 @@ import {MdTooltip, TooltipComponent, MD_TOOLTIP_SCROLL_STRATEGY_PROVIDER} from '
     CommonModule,
     OverlayModule,
     MdCommonModule,
-    PlatformModule
+    PlatformModule,
+    A11yModule,
   ],
   exports: [MdTooltip, TooltipComponent, MdCommonModule],
   declarations: [MdTooltip, TooltipComponent],
