+++ src/lib/tooltip/tooltip.ts
@@ -44,6 +44,7 @@ import {
 } from '@angular/cdk/overlay';
 import {coerceBooleanProperty} from '@angular/cdk/coercion';
 import {ESCAPE} from '@angular/cdk/keyboard';
+import {AriaDescriber} from '@angular/cdk/a11y';
 
 
 export type TooltipPosition = 'left' | 'right' | 'above' | 'below' | 'before' | 'after';
@@ -78,25 +79,6 @@ export const MD_TOOLTIP_SCROLL_STRATEGY_PROVIDER = {
   deps: [Overlay],
   useFactory: MD_TOOLTIP_SCROLL_STRATEGY_PROVIDER_FACTORY
 };
-
-/**
- * Interface used to register tooltip a11y message elements, with a count of how many tooltips have
- * the message and the reference to the a11y message element used for the aria-describedby.
- */
-export interface RegisteredA11yMessage {
-  element: HTMLElement;
-  count: number;
-}
-
-/** ID used for the body container where all a11y messages are appended. */
-export const A11Y_MESSAGES_CONTAINER_ID = 'md-tooltip-a11y-messages';
-
-/** Global incremental identifier for each registered a11y message. */
-let latestA11yMessageId = 0;
-
-/** Global map of all registered a11y message elements that have been placed into the document. */
-const a11yMessages = new Map<string, RegisteredA11yMessage>();
-
 /**
  * Directive that attaches a material design tooltip to the host element. Animates the showing and
  * hiding of a tooltip provided position (defaults to below the element).
@@ -165,12 +147,13 @@ export class MdTooltip implements OnDestroy {
   /** The message to be displayed in the tooltip */
   @Input('mdTooltip') get message() { return this._message; }
   set message(value: string) {
-    if (this._message) { this._unregisterA11yMessage(this._message); }
+    if (this._message) { this._ariaDescriber.unregisterMessage(this._message); }
 
     // If the message is not a string (e.g. number), convert it to a string and trim it.
     this._message = value ? `${value}`.trim() : '';
     this._updateTooltipMessage();
-    this._registerA11yMessage();
+
+    this._setAriaDescribedBy();
   }
 
   /** Classes to be passed to the tooltip. Supports the same syntax as `ngClass`. */
@@ -229,6 +212,7 @@ export class MdTooltip implements OnDestroy {
     private _ngZone: NgZone,
     private _renderer: Renderer2,
     private _platform: Platform,
+    private _ariaDescriber: AriaDescriber,
     @Inject(MD_TOOLTIP_SCROLL_STRATEGY) private _scrollStrategy,
     @Optional() private _dir: Directionality) {
 
@@ -255,7 +239,7 @@ export class MdTooltip implements OnDestroy {
       this._leaveListener();
     }
 
-    this._unregisterA11yMessage(this.message);
+    this._ariaDescriber.unregisterMessage(this.message);
   }
 
   /** Shows the tooltip after the delay in ms, defaults to tooltip-delay-show or 0ms if no input */
@@ -423,80 +407,6 @@ export class MdTooltip implements OnDestroy {
     }
   }
 
-
-  /**
-   * Registers the tooltip message for accessibility. If this message has not yet been
-   * registered, a new element will be created in a visually hidden container with the message
-   * as its content. This provides screenreaders a reference the tooltip trigger's aria-describedby.
-   * If an element has already been created for this message, increase that registered message's
-   * reference count.
-   */
-  private _registerA11yMessage() {
-    if (!this.message) { return; }
-
-    const a11yMessage = a11yMessages.get(this.message);
-    if (a11yMessage) {
-      a11yMessage.count++;
-    } else {
-      this._createA11yMessageElement(this.message);
-    }
-
-    this._setAriaDescribedBy();
-  }
-
-  /**
-   * Removes the a11y tooltip message element if no other tooltips are registered with the same
-   * message. Otherwise, decrease the reference count of the registered a11y tooltip message.
-   */
-  private _unregisterA11yMessage(message: string) {
-    if (!this._platform.isBrowser || !this.message) { return; }
-
-    const a11yMessageElement = a11yMessages.get(message)!;
-    a11yMessageElement.count--;
-
-    // Remove the a11y message if this was its last unique instance.
-    if (a11yMessageElement.count == 0) {
-      this._deleteA11yMessageElement(message);
-    }
-
-    // If the global messages container no longer has any children, remove it.
-    if (!this._getA11yMessagesContainer()!.childNodes.length) {
-      this._renderer.removeChild(document.body, this._getA11yMessagesContainer());
-    }
-
-  }
-
-  /**
-   * Creates a new element in the visually hidden a11y message container element with the message
-   * as its content.
-   */
-  private _createA11yMessageElement(message: string) {
-    if (!this._platform.isBrowser) { return; }
-
-    // Create a visually hidden container for the accessibility messages if one does not yet exist.
-    if (!this._getA11yMessagesContainer()) {
-      this._createA11yMessagesContainer();
-    }
-
-    const messageElement = this._renderer.createElement('div');
-    messageElement.id = `md-tooltip-message-${latestA11yMessageId++}`;
-    messageElement.textContent = message;
-
-    this._getA11yMessagesContainer()!.appendChild(messageElement);
-    a11yMessages.set(message, {element: messageElement, count: 1});
-
-    return messageElement;
-  }
-
-  /** Deletes the a11y message element from the global a11y tooltip messages container. */
-  private _deleteA11yMessageElement(message: string) {
-    if (!this._platform.isBrowser) { return; }
-
-    const a11yMessage = a11yMessages.get(message)!;
-    this._renderer.removeChild(this._getA11yMessagesContainer()!, a11yMessage.element);
-    a11yMessages.delete(message);
-  }
-
   /** Returns the trigger's aria-describedby attribute. */
   private _getAriaDescribedby(): string {
     return this._elementRef.nativeElement.getAttribute('aria-describedby');
@@ -509,25 +419,13 @@ export class MdTooltip implements OnDestroy {
     if (ariaDescribedBy && ariaDescribedBy.indexOf('md-tooltip-message') == -1) { return; }
 
     if (this.message) {
-      const tooltipMessageElementId = a11yMessages.get(this.message)!.element.id;
+      const registeredA11yMessage = this._ariaDescriber.registerMessage(this.message);
+      const tooltipMessageElementId = registeredA11yMessage.element.id;
       this._renderer.setAttribute(
           this._elementRef.nativeElement, 'aria-describedby', tooltipMessageElementId);
     } else {
       this._renderer.setAttribute(this._elementRef.nativeElement, 'aria-describedby', '');
     }
-  }
-
-  /** Creates the global container for all tooltip a11y messages. */
-  private _createA11yMessagesContainer() {
-    const a11yMessagesContainer = this._renderer.createElement('div');
-    a11yMessagesContainer.id = A11Y_MESSAGES_CONTAINER_ID;
-    this._renderer.addClass(a11yMessagesContainer, 'cdk-visually-hidden');
-    this._renderer.appendChild(document.body, a11yMessagesContainer);
-  }
-
-  /** Returns the global container for tooltip a11y messages. */
-  private _getA11yMessagesContainer() {
-    return document.getElementById(A11Y_MESSAGES_CONTAINER_ID);
   }
 }
